fix ft_strlen

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: kadachi <kadachi@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/05 19:44:57 by kadachi           #+#    #+#             */
-/*   Updated: 2024/11/10 19:39:31 by kadachi          ###   ########.fr       */
+/*   Updated: 2024/11/30 18:07:24 by kadachi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
